requirements.txt
asgiref==3.7.2
build==1.2.2.post1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.8
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.54.0
django-bootstrap-calendar==0.1.1
django-crispy-forms==2.0
django-filter==23.2
django-rest-framework==0.1.0
django-scheduler==0.10.1
djangorestframework==3.14.0
icalendar==5.0.7
idna==3.4
jsmin==3.0.1
Markdown==3.4.3
numpy==1.25.0
oauthlib==3.2.2
packaging==24.2
Pillow==10.0.0
pip-tools==7.4.1
psutil==5.9.5
psycopg2-binary==2.9.10
pycparser==2.21
PyJWT==2.8.0
pyproject_hooks==1.2.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tomli==2.2.1
typing_extensions==4.6.3
urllib3==2.0.4
